HTML Assignment
This assignment file provides an opportunity for students to demonstrate their understanding of basic HTML syntax and concepts by fixing an HTML file that is riddled with errors and violations of best practices. The students are encouraged to draw upon concepts presented in the class and use validation tools to help them discover the errors/problems.